backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic[email]
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
celery
redis
alembic
python-multipart